+++ include/roq/order_update.hpp
@@ -22,6 +22,7 @@
 #include "roq/order_status.hpp"
 #include "roq/order_type.hpp"
 #include "roq/position_effect.hpp"
+#include "roq/quantity_type.hpp"
 #include "roq/side.hpp"
 #include "roq/time_in_force.hpp"
 #include "roq/trace.hpp"
@@ -39,6 +40,7 @@ struct ROQ_PUBLIC OrderUpdate final {
   roq::Side side = {};                                          //!< Side
   roq::PositionEffect position_effect = {};                     //!< Position effect
   roq::MarginMode margin_mode = {};                             //!< Margin mode
+  roq::QuantityType quantity_type = {};                         //!< Type of quantity (requires ecxhange support)
   double max_show_quantity = roq::NaN;                          //!< Quantity visible to market (base currency, requires exchange support)
   roq::OrderType order_type = {};                               //!< Order type
   roq::TimeInForce time_in_force = {};                          //!< Time in force
@@ -94,6 +96,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         R"(side={}, )"
         R"(position_effect={}, )"
         R"(margin_mode={}, )"
+        R"(quantity_type={}, )"
         R"(max_show_quantity={}, )"
         R"(order_type={}, )"
         R"(time_in_force={}, )"
@@ -132,6 +135,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         value.side,
         value.position_effect,
         value.margin_mode,
+        value.quantity_type,
         value.max_show_quantity,
         value.order_type,
         value.time_in_force,