+++ include/roq/reference_data.hpp
@@ -30,6 +30,7 @@ struct ROQ_PUBLIC ReferenceData final {
   roq::SecurityType security_type = {};             //!< Security type
   std::string_view base_currency;                   //!< Base currency
   std::string_view quote_currency;                  //!< Quote currency
+  std::string_view settlement_currency;             //!< Settlement currency
   std::string_view margin_currency;                 //!< Margin currency
   std::string_view commission_currency;             //!< Commission currency
   double tick_size = roq::NaN;                      //!< Minimum price increment
@@ -76,6 +77,7 @@ struct fmt::formatter<roq::ReferenceData> {
         R"(security_type={}, )"
         R"(base_currency="{}", )"
         R"(quote_currency="{}", )"
+        R"(settlement_currency="{}", )"
         R"(margin_currency="{}", )"
         R"(commission_currency="{}", )"
         R"(tick_size={}, )"
@@ -105,6 +107,7 @@ struct fmt::formatter<roq::ReferenceData> {
         value.security_type,
         value.base_currency,
         value.quote_currency,
+        value.settlement_currency,
         value.margin_currency,
         value.commission_currency,
         value.tick_size,